+++ example/genkit_example.dart
@@ -1,4 +1,4 @@
-import 'package:genkit/genkit.dart';
+import 'package:genkit/client.dart';
 import 'package:http/http.dart' as http;
 
 import '../test/schemas/my_schemas.dart';
@@ -6,24 +6,81 @@ import '../test/schemas/stream_schemas.dart';
 
 const baseUrl = 'http://localhost:8080';
 
-// Example 1: A simple flow that takes a string and returns a string.
-Future<void> _runStringFlow(http.Client client) async {
-  print('--- 1. String to String flow ---');
+void printServerInstructions() {
+  print(
+    '-------------------------------------------------------------------\n'
+    '| Before running these examples, make sure the server is running. |\n'
+    '| The server is using a fake model, so does not require API keys. |\n'
+    '| In a separate terminal run:                                     |\n'
+    '|                                                                 |\n'
+    '| \$ cd example/server                                             |\n'
+    '| \$ npm i                                                         |\n'
+    '| \$ npm start                                                     |\n'
+    '-------------------------------------------------------------------\n'
+  );
+}
+
+// A simple flow that takes a string and returns a string.
+Future<void> _runStringFlow() async {
+  print('--- String to String flow ---');
   final echoStringFlow = defineRemoteAction(
     url: '$baseUrl/echoString',
-    httpClient: client,
     fromResponse: (json) => json as String,
   );
   final response = await echoStringFlow(input: 'Hello Genkit client for Dart!');
   print('Response: $response');
 }
 
-// Example 2: A flow that takes an object and returns an object.
-Future<void> _runObjectFlow(http.Client client) async {
-  print('\n--- 2. Object to Object flow ---');
+// Error handling when calling remote flows.
+Future<void> _runThrowingFlow() async {
+  print('\n--- Flow error handling ---');
+  final throwy = defineRemoteAction(
+    url: '$baseUrl/throwy',
+    fromResponse: (json) => json as String,
+  );
+  try {
+    await throwy(input: 'Hello Genkit client for Dart!');
+  } on GenkitException catch (e) {
+    if (e.underlyingException is http.ClientException) {
+      print('Client error: ${e.underlyingException}');
+      print('Make sure the server is running.');
+    } else {
+      print('Excepted flow error: ${e.details}');
+    }
+  }
+}
+
+// Error handling when calling remote flows.
+Future<void> _runThrowingStreamingFlow() async {
+  print('\n--- Streaming Flow error handling ---');
+  final streamyThrowy = defineRemoteAction(
+    url: '$baseUrl/streamyThrowy',
+    fromResponse: (json) => json as String,
+    fromStreamChunk: (json) => json,
+  );
+  try {
+    final (:stream, :response) = streamyThrowy.stream(input: 5);
+
+    await for (final chunk in stream) {
+      print('Chunk: $chunk');
+    }
+  } on GenkitException catch (e) {
+    if (e.underlyingException is http.ClientException) {
+      print('Client error: ${e.underlyingException}');
+      print('Make sure the server is running.');
+    } else {
+      print('Excepted flow error: ${e.details}');
+    }
+  } catch (e, st) {
+    print('Caught error: $e $st');
+  }
+}
+
+// A flow that takes an object and returns an object.
+Future<void> _runObjectFlow() async {
+  print('\n--- Object to Object flow ---');
   final processObjectFlow = defineRemoteAction(
     url: '$baseUrl/processObject',
-    httpClient: client,
     fromResponse: (json) => MyOutput.fromJson(json),
   );
   final response = await processObjectFlow(
@@ -32,12 +89,11 @@ Future<void> _runObjectFlow(http.Client client) async {
   print('Response: ${response.reply}');
 }
 
-// Example 3: A streaming flow.
-Future<void> _runStreamingFlow(http.Client client) async {
-  print('\n--- 3. Stream generate call ---');
+// A streaming flow.
+Future<void> _runStreamingFlow() async {
+  print('\n--- Stream generate call ---');
   final streamObjectsFlow = defineRemoteAction(
     url: '$baseUrl/streamObjects',
-    httpClient: client,
     fromResponse: (json) => StreamOutput.fromJson(json),
     fromStreamChunk: (json) => StreamOutput.fromJson(json),
   );
@@ -51,12 +107,12 @@ Future<void> _runStreamingFlow(http.Client client) async {
   }
   print('\nStream finished.');
   final finalResult = await response;
-  print('Final Response: ${finalResult.text}');
+  print('Final Response: ${finalResult?.text}');
 }
 
-// --- 4. Stream generate call ---
-Future<void> _runStreamingGenerateFlow(http.Client client) async {
-  print('\n--- 4. Stream generate call ---');
+// --- Stream generate call ---
+Future<void> _runStreamingGenerateFlow() async {
+  print('\n--- Stream generate call ---');
   final generateFlow = defineRemoteAction(
     url: '$baseUrl/generate',
     fromResponse: (json) => GenerateResponse.fromJson(json),
@@ -85,37 +141,42 @@ Future<void> _runStreamingGenerateFlow(http.Client client) async {
   }
   print('\nStream finished.');
   final finalResult = await response;
-  print('Final Response: ${finalResult.text}');
+  print('Final Response: ${finalResult?.text}');
 }
 
-// Example 4: Manual client management for performance.
+// Manual client management for performance.
 // Reusing a single client for multiple calls is more efficient.
-Future<void> _runPerformanceExample(http.Client client) async {
-  print('\n--- 5. Manual Client Management for Performance ---');
-  final echoAction = defineRemoteAction(
-    url: '$baseUrl/echoString',
-    httpClient: client,
-    fromResponse: (json) => json as String,
-  );
+Future<void> _runPerformanceExample() async {
+  final client = http.Client();
+  try {
+    print('\n--- Manual Client Management for Performance ---');
+    final echoAction = defineRemoteAction(
+      url: '$baseUrl/echoString',
+      httpClient: client,
+      fromResponse: (json) => json as String,
+    );
 
-  final r1 = await echoAction(input: 'First call');
-  print('First response: $r1');
-  final r2 = await echoAction(input: 'Second call');
-  print('Second response: $r2');
+    final r1 = await echoAction(input: 'First call');
+    print('First response: $r1');
+    final r2 = await echoAction(input: 'Second call');
+    print('Second response: $r2');
+  } finally {
+    print('\nClosing HTTP client.');
+    client.close();
+  }
 }
 
 void main() async {
-  // It's recommended to create a single client and reuse it for all requests.
-  final client = http.Client();
   try {
-    await _runStringFlow(client);
-    await _runObjectFlow(client);
-    await _runStreamingFlow(client);
-    await _runStreamingGenerateFlow(client);
-    await _runPerformanceExample(client);
-  } finally {
-    // Ensure the client is closed when all operations are complete.
-    print('\nClosing HTTP client.');
-    client.close();
+    await _runStringFlow();
+    await _runObjectFlow();
+    await _runStreamingFlow();
+    await _runStreamingGenerateFlow();
+    await _runPerformanceExample();
+    await _runThrowingFlow();
+    await _runThrowingStreamingFlow();
+  } catch (e, st) {
+    print('$e\n$st\n');
+    printServerInstructions();
   }
 }