+++ example/server/index.ts
@@ -1,4 +1,4 @@
-import { genkit, z } from "genkit";
+import { genkit, UserFacingError, z } from "genkit";
 import {
   GenerateResponseChunkSchema,
   GenerateResponseSchema,
@@ -15,16 +15,12 @@ ai.defineModel(
     name: "echoModel",
   },
   async (request, { sendChunk }) => {
-    sendChunk({
-      content: [{ text: "chunk 1" }],
-    });
-    sendChunk({
-      content: [{ text: "chunk 2" }],
-    });
-    sendChunk({
-      content: [{ text: "chunk 3" }],
-    });
-
+    for (var i = 0; i < 3; i++) {
+      sendChunk({
+        content: [{ text: `chunk ${i}` }],
+      });
+      await new Promise((r) => setTimeout(r, 1000));
+    }
     return {
       message: {
         role: "model",
@@ -87,6 +83,7 @@ const streamObjects = ai.defineFlow(
         text: "input: " + i,
         summary: "summary " + i,
       });
+      await new Promise((r) => setTimeout(r, 1000));
     }
 
     return {
@@ -110,6 +107,41 @@ const generate = ai.defineFlow(
     })
 );
 
+export const streamyThrowy = ai.defineFlow(
+  {
+    name: "streamyThrowy",
+    inputSchema: z.number(),
+    outputSchema: z.string(),
+    streamSchema: z.object({ count: z.number() }),
+  },
+  async (count, { sendChunk }) => {
+    let i = 0;
+    for (; i < count; i++) {
+      if (i == 3) {
+        throw new UserFacingError("INTERNAL", "whoops");
+      }
+      await new Promise((r) => setTimeout(r, 1000));
+      sendChunk({ count: i });
+    }
+    return `done: ${count}, streamed: ${i} times`;
+  }
+);
+
+export const throwy = ai.defineFlow(
+  { name: "throwy", inputSchema: z.string(), outputSchema: z.string() },
+  async (subject) => {
+    const foo = await ai.run("call-llm", async () => {
+      return `subject: ${subject}`;
+    });
+    if (subject) {
+      throw new UserFacingError("INTERNAL", "whoops");
+    }
+    return await ai.run("call-llm", async () => {
+      return `foo: ${foo}`;
+    });
+  }
+);
+
 const app = express();
 app.use(express.json());
 
@@ -117,5 +149,7 @@ app.post("/echoString", expressHandler(echoString));
 app.post("/generate", expressHandler(generate));
 app.post("/processObject", expressHandler(processObject));
 app.post("/streamObjects", expressHandler(streamObjects));
+app.post("/throwy", expressHandler(throwy));
+app.post("/streamyThrowy", expressHandler(streamyThrowy));
 
 app.listen(8080);