+++ test/client_test.dart
@@ -15,7 +15,7 @@
 import 'dart:convert';
 import 'dart:async';
 
-import 'package:genkit/genkit.dart';
+import 'package:genkit/client.dart';
 import 'package:test/test.dart';
 import 'package:http/http.dart' as http;
 import 'package:mockito/annotations.dart';